chore: move types to root src dir
