+++ sdk-js/src/cohere/cohere.ts
@@ -17,7 +17,7 @@ import {
   PartialModelResponse,
   Tool,
   ToolCallPart,
-} from "../schema/index.js";
+} from "../types.js";
 import { convertAudioPartsToTextParts } from "../utils/message.utils.js";
 import { ContentDeltaAccumulator } from "../utils/stream.utils.js";
 import { calculateCost } from "../utils/usage.utils.js";
