+++ sdk-js/src/mistral/mistral.ts
@@ -18,7 +18,7 @@ import {
   PartialModelResponse,
   Tool,
   ToolCallPart,
-} from "../schema/index.js";
+} from "../types.js";
 import { convertAudioPartsToTextParts } from "../utils/message.utils.js";
 import {
   ContentDeltaAccumulator,
