+++ sdk-js/src/types.ts
@@ -1,8 +1,5 @@
 /**
  * The encoding of the audio.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "AudioEncoding".
  */
 export type AudioEncoding =
   | "linear16"
@@ -14,30 +11,15 @@ export type AudioEncoding =
   | "opus";
 /**
  * The container format of the audio.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "AudioContainer".
  */
 export type AudioContainer = "wav" | "ogg" | "flac" | "webm";
-/**
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "Part".
- */
 export type Part =
   | TextPart
   | ImagePart
   | AudioPart
   | ToolCallPart
   | ToolResultPart;
-/**
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "Message".
- */
 export type Message = UserMessage | AssistantMessage | ToolMessage;
-/**
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "Modality".
- */
 export type Modality = "text" | "audio";
 
 export interface LlmSdk {
@@ -45,9 +27,6 @@ export interface LlmSdk {
 }
 /**
  * A part of the message that contains text.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "TextPart".
  */
 export interface TextPart {
   type: "text";
@@ -59,9 +38,6 @@ export interface TextPart {
 }
 /**
  * A part of the message that contains an image.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "ImagePart".
  */
 export interface ImagePart {
   type: "image";
@@ -88,9 +64,6 @@ export interface ImagePart {
 }
 /**
  * A part of the message that contains an audio.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "AudioPart".
  */
 export interface AudioPart {
   type: "audio";
@@ -119,9 +92,6 @@ export interface AudioPart {
 }
 /**
  * A part of the message that represents a call to a tool the model wants to use.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "ToolCallPart".
  */
 export interface ToolCallPart {
   type: "tool-call";
@@ -146,9 +116,6 @@ export interface ToolCallPart {
 }
 /**
  * A part of the message that represents the result of a tool call.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "ToolResultPart".
  */
 export interface ToolResultPart {
   type: "tool-result";
@@ -175,9 +142,6 @@ export interface ToolResultPart {
 }
 /**
  * Represents a message sent by the user.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "UserMessage".
  */
 export interface UserMessage {
   role: "user";
@@ -185,18 +149,11 @@ export interface UserMessage {
 }
 /**
  * Represents a message generated by the model.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "AssistantMessage".
  */
 export interface AssistantMessage {
   role: "assistant";
   content: (TextPart | ToolCallPart | AudioPart)[];
 }
-/**
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "TextPartDelta".
- */
 export interface TextPartDelta {
   type: "text";
   text: string;
@@ -205,10 +162,6 @@ export interface TextPartDelta {
    */
   id?: string;
 }
-/**
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "ToolCallPartDelta".
- */
 export interface ToolCallPartDelta {
   type: "tool-call";
   /**
@@ -228,10 +181,6 @@ export interface ToolCallPartDelta {
    */
   id?: string;
 }
-/**
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "AudioPartDelta".
- */
 export interface AudioPartDelta {
   type: "audio";
   /**
@@ -257,28 +206,18 @@ export interface AudioPartDelta {
    */
   id?: string;
 }
-/**
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "ContentDelta".
- */
 export interface ContentDelta {
   index: number;
   part: TextPartDelta | ToolCallPartDelta | AudioPartDelta;
 }
 /**
  * Represents a JSON schema.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "JSONSchema".
  */
 export interface JSONSchema {
   [k: string]: unknown;
 }
 /**
  * Represents a tool that can be used by the model.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "Tool".
  */
 export interface Tool {
   /**
@@ -296,9 +235,6 @@ export interface Tool {
 }
 /**
  * Represents tool result in the message history.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "ToolMessage".
  */
 export interface ToolMessage {
   role: "tool";
@@ -306,9 +242,6 @@ export interface ToolMessage {
 }
 /**
  * Represents the token usage of the model.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "ModelTokensDetail".
  */
 export interface ModelTokensDetail {
   textTokens?: number;
@@ -317,9 +250,6 @@ export interface ModelTokensDetail {
 }
 /**
  * Represents the token usage of the model.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "ModelUsage".
  */
 export interface ModelUsage {
   inputTokens: number;
@@ -329,9 +259,6 @@ export interface ModelUsage {
 }
 /**
  * Represents the response generated by the model.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "ModelResponse".
  */
 export interface ModelResponse {
   content: (TextPart | ToolCallPart | AudioPart)[];
@@ -341,69 +268,41 @@ export interface ModelResponse {
    */
   cost?: number;
 }
-/**
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "PartialModelResponse".
- */
 export interface PartialModelResponse {
   delta: ContentDelta;
 }
 /**
  * The model will automatically choose the tool to use or not use any tools.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "ToolChoiceAuto".
  */
 export interface ToolChoiceAuto {
   type: "auto";
 }
 /**
  * The model will not use any tools.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "ToolChoiceNone".
  */
 export interface ToolChoiceNone {
   type: "none";
 }
 /**
  * The model will be forced to use a tool.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "ToolChoiceRequired".
  */
 export interface ToolChoiceRequired {
   type: "required";
 }
 /**
  * The model will use the specified tool.
- *
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "ToolChoiceTool".
  */
 export interface ToolChoiceTool {
   type: "tool";
   toolName: string;
 }
-/**
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "ResponseFormatText".
- */
 export interface ResponseFormatText {
   type: "text";
 }
-/**
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "ResponseFormatJson".
- */
 export interface ResponseFormatJson {
   type: "json";
   schema?: JSONSchema;
 }
-/**
- * This interface was referenced by `LlmSdk`'s JSON-Schema
- * via the `definition` "LanguageModelInput".
- */
 export interface LanguageModelInput {
   /**
    * A system prompt is a way of providing context and instructions to the model
