+++ sdk-js/tsconfig.build.json
@@ -3,7 +3,8 @@
   "compilerOptions": {
     "rootDir": "src",
     "outDir": "dist",
-    "declaration": true
+    "declaration": true,
+    "noEmit": false
   },
   "include": ["src"]
 }
